requirements.txt
autogluon.tabular==1.1.1
autogluon.multimodal==1.1.1
mlflow==2.14.2
pandas
pyyaml
